Update cltool_updateFlaghCfg to use a single CltoolDevice, instead of InertialSense class. Update ClToolDevice - public ISDevice and update allocateDevice signature. Misc cleanup in DeviceFactory for cltool, and tests. Additional options work in DeviceManager to work with tests. Misc cleanup in PortManager.

+++ src/serialPort.c
@@ -341,7 +341,7 @@ int serialPortWrite(port_handle_t port, const unsigned char* buffer, unsigned in
 int serialPortWriteLine(port_handle_t port, const unsigned char* buffer, unsigned int writeCount)
 {
     int count = serialPortWrite(port, buffer, writeCount);
-    if (count == writeCount)
+    if (count == (int)writeCount)
         count += serialPortWrite(port, (unsigned char*)"\r\n", 2);
     return count;
 }
